% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vc_df}
\alias{vc_df}
\title{Convenience function to turn vfb_connect summary results into data.frames}
\usage{
vc_df(x, ..., tibble = FALSE)
}
\arguments{
\item{x}{A summary result from a \code{VfbConnect} call}

\item{...}{additional arguments passed to \code{\link{as.data.frame}} or
\code{\link[tibble]{as_tibble}}}

\item{tibble}{Whether to return a \code{\link{tibble}} rather than a regular
data.frame}
}
\description{
Convenience function to turn vfb_connect summary results into data.frames
}
\examples{
vc=VfbConnect()
KC_types = vc$get_subclasses("Kenyon cell", summary=TRUE)
vc_df(KC_types)

vc$neo_query_wrapper$get_type_TermInfo(list('FBbt_00003686'), summary=TRUE) \%>\%
vc_df
}
