% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{vc_install}
\alias{vc_install}
\title{Install or update the VFB connect python module}
\usage{
vc_install(package = "vfb-connect", ...)
}
\arguments{
\item{package}{Optional package name, possible with version specifier (see
examples)}

\item{...}{Additional arguments passed to
\code{reticulate::\link{py_install}}}
}
\description{
This uses \code{reticulate::\link{py_install}} to install or
  update the underlying python package.
}
\examples{
\dontrun{
vc_install()

# install a specific version
vc_install("vfb-connect~=1.2.4")
}
}
