% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vfb_datasets}
\alias{vfb_datasets}
\title{Return summary of datasets available from VFB}
\usage{
vfb_datasets()
}
\value{
A data.frame with columns including DOI
}
\description{
Return summary of datasets available from VFB
}
\examples{
\donttest{
df=vfb_datasets()
head(df)
subset(df, !nzchar(DOI))
}
}
