% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfbconnectr-package.R
\docType{package}
\name{vfbconnectr-package}
\alias{vfbconnectr}
\alias{vfbconnectr-package}
\title{vfbconnectr: An R Interface to vfb-connect to query virtualflybrain.org}
\description{
Provides an R interface to the Python package vfb-connect.
    'VFB_connect' is a Python lib that wraps data / knowledgebase query 
    endpoints and returns 'VFB_json'.
}
\examples{
\dontrun{
# make a VfbConnect to access the Python module
vc=VfbConnect()
}
}
\seealso{
\code{\link{VfbConnect}}
}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
